requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langgraph==0.2.0
pymilvus==2.4.0
psycopg2-binary==2.9.0
python-dotenv==1.0.0
fastapi==0.124.2
uvicorn==0.38.0